➕ ADDS: DataSourceIntegration
