+++ integrations/integration.php
@@ -67,30 +67,12 @@ abstract class Integration {
 	protected bool $have_multiple_setups = false;
 
 	/**
-	 * When an integration doesn't have its own config and is dependent on configs
-	 * of other integrations then we can use this property.
-	 *
-	 * Example:
-	 *
-	 * VIP Composable Blocks integration needs data sources which
-	 * are dependent on Shopify, Airtable etc.
-	 *
-	 * @var array<Integration>
-	 */
-	protected array $child_integrations = [];
-
-	/**
 	 * Constructor.
 	 *
 	 * @param string $slug Slug of the integration.
 	 */
 	public function __construct( string $slug ) {
 		$this->slug = $slug;
-
-		// Registers child integrations if any.
-		foreach ( $this->child_integrations as $integration ) {
-			IntegrationsSingleton::instance()->register( $integration );
-		}
 
 		add_action( 'switch_blog', array( $this, 'switch_blog_callback' ), 10 );
 	}
@@ -150,13 +132,6 @@ abstract class Integration {
 	 * @private
 	 */
 	public function is_active_via_vip(): bool {
-		// Returns 'true` if any of the child integrations is active else get the related config for the integration.
-		foreach ( $this->child_integrations as $integration ) {
-			if ( $integration->is_active_via_vip() ) {
-				return true;
-			}
-		}
-
 		return in_array( Env_Integration_Status::ENABLED, $this->get_site_statuses() );
 	}
 
@@ -234,15 +209,6 @@ abstract class Integration {
 		 */
 		$configs = [];
 
-		// If integration have child integrations then merge the config of each child integration and return.
-		if ( count( $this->child_integrations ) ) {
-			foreach ( $this->child_integrations as $integration ) {
-				$configs = array_merge( $configs, $integration->get_config() );
-			}
-
-			return $configs;
-		}
-
 		// Get configs of the integration from configurations provided by VIP.
 		foreach ( $this->vip_configs as $vip_config ) {
 			if ( is_multisite() ) {
@@ -256,14 +222,14 @@ abstract class Integration {
 				$config = $this->get_value_from_config( $vip_config, 'env', 'config' );
 			}
 
-			if ( isset( $config ) ) {
-				$configs[] = $config;
-			}
-
-			$config['type'] = $this->slug; // Useful to have it available, specially when integration is dependent on child integrations.
-
 			if ( isset( $vip_config['label'] ) ) {
 				$config['label'] = $vip_config['label']; // Useful to differentiate between multiple configs of same integration.
+			}
+
+			if ( isset( $config ) ) {
+				$config['type'] = $this->slug; // Useful to have it available, specially when integration is dependent on multiple integrations.
+
+				$configs[] = $config;
 			}
 		}
